+++ src/components/icons/siteLogo.tsx
@@ -1,8 +1,9 @@
 interface SvgProps extends React.SVGProps<SVGSVGElement> {
   className?: string;
+  color?: string;
 }
 
-const SiteLogo = ({ className }: SvgProps) => (
+const SiteLogo = ({ className, color = "#F8FAFC" }: SvgProps) => (
   <svg
     xmlns="http://www.w3.org/2000/svg"
     viewBox="0 0 120 25"
@@ -10,7 +11,7 @@ const SiteLogo = ({ className }: SvgProps) => (
     fill="none"
   >
     <path
-      fill="#F8FAFC"
+      fill={color}
       d="M103.291 4.296h2.097c.85-.016 1.744-.049 2.68-.098a28.062 28.062 0 0 0 2.594-.292 23.707 23.707 0 0 0 1.6-.293 63.83 63.83 0 0 0 1.513-.341c.49-.13.929-.268 1.318-.415.404-.146.742-.284 1.016-.414l1.773 2.682c-.375.097-.699.187-.973.268-.274.081-.504.146-.692.195-.446.13-.929.26-1.448.39a41.4 41.4 0 0 1-1.599.317c-.548.098-1.103.195-1.665.293-.821.13-1.693.227-2.615.292-.923.065-1.816.122-2.681.171-.864.033-1.613.057-2.248.073l-.67-2.828Zm.93 16.8c1.268-.651 2.305-1.44 3.112-2.366a9.074 9.074 0 0 0 1.816-3.243c.404-1.235.605-2.617.605-4.145V6.32l2.832-.366v5.34c0 1.625-.18 3.17-.54 4.632a12.415 12.415 0 0 1-1.751 4.048c-.822 1.219-1.931 2.3-3.329 3.242l-2.745-2.12Zm-2.551-10.973c.303.033.627.065.973.098.346.032.684.049 1.016.049h14.374c.158 0 .36-.009.605-.025l.735-.049c.259-.032.468-.056.627-.073v3.146a19.422 19.422 0 0 0-1.902-.098h-14.396c-.331 0-.684.008-1.059.024-.375.017-.699.041-.973.073v-3.145ZM87.636 9.775c.162.02.37.052.621.094l.783.062c.279.02.535.031.769.031h7.246c.315 0 .612-.02.89-.062.28-.063.51-.104.69-.125v4.888l-.69-.093a6.606 6.606 0 0 0-.89-.063H89.81c-.387 0-.783.021-1.188.063-.404.02-.733.051-.985.093V9.775ZM84.566 3.666c-.016.313-.04.656-.072 1.03-.016.357-.024.76-.024 1.208v4.431c0 1.193-.064 2.238-.191 3.133-.112.895-.287 1.678-.526 2.35a8.024 8.024 0 0 1-.885 1.79c-.335.507-.726.992-1.172 1.455a10.004 10.004 0 0 1-1.818 1.454c-.686.418-1.38.769-2.081 1.052-.686.298-1.324.537-1.914.716l-2.511-2.462c1.195-.253 2.272-.604 3.229-1.051a9.06 9.06 0 0 0 2.535-1.768c.415-.433.75-.866 1.005-1.298.255-.433.446-.91.574-1.432.143-.537.239-1.142.287-1.813.064-.671.096-1.44.096-2.305V5.904c0-.448-.016-.85-.048-1.209a9.546 9.546 0 0 0-.096-1.029h3.612Zm-11.099.18-.048.805a14.9 14.9 0 0 0-.023.917v7.564c0 .284.007.612.023.985s.032.694.048.962H70l.072-.895c.032-.388.047-.746.047-1.074V5.568c0-.209-.007-.477-.023-.805-.016-.344-.04-.65-.072-.918h3.444ZM52.495 4.89c.396.014.766.029 1.11.043h12.313l1.033-.044-.038 1.073v12.603c.013.321.02.65.02.985v.853l.018.372H64.31v-.525c.013-.336.02-.73.02-1.182.012-.452.018-.875.018-1.269V7.712H55.1v11.925c.013.307.02.562.02.766v.372h-2.643v-.372c.013-.219.019-.489.019-.81.013-.32.02-.649.02-.984.012-.336.018-.635.018-.897V5.96c0-.408-.013-.765-.038-1.072Zm12.81 11.903v2.8H53.797v-2.8h11.508ZM41.3 2.452c-.032.253-.064.582-.096.984-.016.403-.032.739-.047 1.007-.064 2.223-.264 4.2-.598 5.93-.335 1.716-.814 3.238-1.436 4.566a14.448 14.448 0 0 1-2.368 3.58c-.957 1.045-2.097 2.007-3.42 2.888l-2.775-2.104c.495-.269 1.02-.582 1.579-.94a9.538 9.538 0 0 0 1.555-1.253c.781-.716 1.443-1.5 1.985-2.35.542-.865.973-1.805 1.292-2.82.335-1.03.582-2.156.741-3.379.16-1.223.24-2.573.24-4.05 0-.165-.009-.373-.025-.627-.016-.269-.04-.53-.071-.783a5.886 5.886 0 0 0-.072-.65H41.3Zm8.133 4.789c-.032.179-.072.388-.12.626a4.502 4.502 0 0 0-.048.56 62.23 62.23 0 0 1-.096 1.835 49.548 49.548 0 0 1-.167 2.417c-.064.85-.151 1.7-.263 2.551-.096.85-.215 1.641-.359 2.372-.143.716-.319 1.305-.526 1.768-.255.567-.63.992-1.124 1.276-.479.298-1.133.447-1.962.447-.685 0-1.371-.022-2.057-.067a60.074 60.074 0 0 1-1.913-.112l-.36-2.999c.59.105 1.18.187 1.771.247.59.044 1.116.067 1.579.067.398 0 .693-.06.885-.18.191-.134.35-.335.478-.604a7.09 7.09 0 0 0 .359-1.186 18.14 18.14 0 0 0 .287-1.656c.08-.611.143-1.245.191-1.902.064-.656.104-1.29.12-1.902.032-.627.048-1.179.048-1.656H34.483c-.399 0-.893.007-1.483.022-.59 0-1.148.023-1.674.067V6.301c.51.045 1.052.082 1.626.112.574.03 1.084.044 1.531.044h11.075c.303 0 .606-.014.909-.044.319-.03.614-.067.885-.112l2.08.94Z"
     />
     <rect
